Refactored Project structure, reworked relative path in test class
