+++ SecretSanta/SecretSanta.sln
@@ -11,14 +11,14 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{1768D2E7-9
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Domain.Tests", "test\SecretSanta.Domain.Tests\SecretSanta.Domain.Tests.csproj", "{BC742C5A-001D-4792-B767-45DE135011CD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Library", "SecretSanta.Library\SecretSanta.Library.csproj", "{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Library.Tests", "SecretSanta.Library.Tests\SecretSanta.Library.Tests.csproj", "{923FF742-C566-4B76-9B21-741ECF21178D}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TestData", "TestData", "{8CFA0872-BD75-4B2B-9B18-E5B5F2B1DFFC}"
 	ProjectSection(SolutionItems) = preProject
 		ValidFile.txt = ValidFile.txt
 	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Library", "src\SecretSanta.Library\SecretSanta.Library.csproj", "{C0872F70-6662-4CC3-997C-0537A5C15800}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Library.Tests", "test\SecretSanta.Library.Tests\SecretSanta.Library.Tests.csproj", "{F96F9CA3-2EAC-4E68-805C-D9C09795D069}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -54,30 +54,30 @@ Global
 		{BC742C5A-001D-4792-B767-45DE135011CD}.Release|x64.Build.0 = Release|Any CPU
 		{BC742C5A-001D-4792-B767-45DE135011CD}.Release|x86.ActiveCfg = Release|Any CPU
 		{BC742C5A-001D-4792-B767-45DE135011CD}.Release|x86.Build.0 = Release|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Debug|x64.Build.0 = Debug|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Debug|x86.Build.0 = Debug|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Release|x64.ActiveCfg = Release|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Release|x64.Build.0 = Release|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Release|x86.ActiveCfg = Release|Any CPU
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0}.Release|x86.Build.0 = Release|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Debug|x64.Build.0 = Debug|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Debug|x86.Build.0 = Debug|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Release|x64.ActiveCfg = Release|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Release|x64.Build.0 = Release|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Release|x86.ActiveCfg = Release|Any CPU
-		{923FF742-C566-4B76-9B21-741ECF21178D}.Release|x86.Build.0 = Release|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Debug|x64.Build.0 = Debug|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Debug|x86.Build.0 = Debug|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Release|x64.ActiveCfg = Release|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Release|x64.Build.0 = Release|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Release|x86.ActiveCfg = Release|Any CPU
+		{C0872F70-6662-4CC3-997C-0537A5C15800}.Release|x86.Build.0 = Release|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Debug|x64.Build.0 = Debug|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Debug|x86.Build.0 = Debug|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Release|x64.ActiveCfg = Release|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Release|x64.Build.0 = Release|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Release|x86.ActiveCfg = Release|Any CPU
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -85,9 +85,9 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{C338FD95-38B3-4B1D-8864-E9314D46C0EE} = {EE6EE349-1578-464C-ACA6-8E537A823075}
 		{BC742C5A-001D-4792-B767-45DE135011CD} = {1768D2E7-9C81-4FB2-B22D-C8B09767B9B9}
-		{33E6C340-F4A9-440D-8A9A-4A01DB9F65B0} = {EE6EE349-1578-464C-ACA6-8E537A823075}
-		{923FF742-C566-4B76-9B21-741ECF21178D} = {1768D2E7-9C81-4FB2-B22D-C8B09767B9B9}
 		{8CFA0872-BD75-4B2B-9B18-E5B5F2B1DFFC} = {1768D2E7-9C81-4FB2-B22D-C8B09767B9B9}
+		{C0872F70-6662-4CC3-997C-0537A5C15800} = {EE6EE349-1578-464C-ACA6-8E537A823075}
+		{F96F9CA3-2EAC-4E68-805C-D9C09795D069} = {1768D2E7-9C81-4FB2-B22D-C8B09767B9B9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B800A818-3643-4782-BF10-DAA27D9DE733}
