+++ SecretSanta/src/SecretSanta.Library/SecretSanta.Library.csproj
