+++ SecretSanta/test/SecretSanta.Library.Tests/SecretSanta.Library.Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.2</TargetFramework>
@@ -13,7 +13,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\SecretSanta.Library\SecretSanta.Library.csproj" />
+    <ProjectReference Include="..\..\src\SecretSanta.Library\SecretSanta.Library.csproj" />
   </ItemGroup>
+
 
 </Project>
